- apply location,contacts & bank via UI #173 - rename bpartner utils
